refactor(ios): extract shared ChatContentView to eliminate duplicate chat UI
